Number of Digits and Print Permutations
Assignment 1

numberOfDigits is a function that should take one unsigned long long int as input and return a unsigned long long int value. numberOfDigits should return the number of digits in the input decimal integer.

printPermutations is a function that should take one string and one integer n as input and print out the first n unique permutations of that string using the Johnson-Trotter algorithm. Each permutation should appear on a separate line. You can assume that the string is already in lexicographic order.